Fixing the sign in bug, attempt one.
This tries to fix the sign-in character loading bug by pausing the character loading code, so that it cannot run ahead of itself.  Testing in production required, the sleep in the new_player.dm makes it so that it doesn't present a random character (despite actually having loaded it anyways).
:cl:
code: Character and pref loading delay, attempting to not let the code run ahead of itself.
/:cl: